+++ hvega/hvega.nix
@@ -1,5 +1,6 @@
 { mkDerivation, aeson, aeson-pretty, base, bytestring, containers
 , filepath, stdenv, tasty, tasty-golden, text, unordered-containers
+, zlib
 }:
 mkDerivation {
   pname = "hvega";
@@ -8,6 +9,7 @@ mkDerivation {
   isLibrary = true;
   isExecutable = true;
   libraryHaskellDepends = [ aeson base text unordered-containers ];
+  executableHaskellDepends = [ zlib ];
   testHaskellDepends = [
     aeson aeson-pretty base bytestring containers filepath tasty
     tasty-golden text