+++ hvega/CHANGELOG.md
@@ -21,7 +21,15 @@ have been changed to accept an additional argument (the new
 configuration should be applied to. This is to support new axis
 configuration options added in Vega-Lite 4.7.0.
 
+The `ChTooltip` `Channel` constructor has been removed as support
+for this channel type was dropped in Vega-Lite 4.
+
 ### New constructors
+
+The `ScaleDomain` type has gained `DSelectionField` and
+`DSelectionChannel` constructors, which allow you to link a scale
+(e.g. an axis) to a selection that is projected over multiple fields
+or encodings.
 
 The `Operation` type has gained the `Product` specifier from Vega-Lite
 4.6.0.